Add analysis runner, report generator, and fix float32 JSON bug
- run_analysis.py: runs full Q1 analysis on all word families using synthetic
  embedding models; generates 4 plots + JSON results to q1_report/
  (swap SyntheticModel for real KeyedVectors once AraVec / GloVe downloaded)
- generate_report.py: standalone illustrative report with Markdown summary
- Fix float32 not JSON serializable: cast cosine_similarity return to float()
- Update .gitignore: exclude venv/ and q1_report/ output directory

+++ q1_root_cluster_density.py
@@ -168,7 +168,7 @@ def cosine_similarity(v1, v2):
     norm2 = np.linalg.norm(v2)
     if norm1 == 0 or norm2 == 0:
         return 0.0
-    return np.dot(v1, v2) / (norm1 * norm2)
+    return float(np.dot(v1, v2) / (norm1 * norm2))
 
 
 def compute_pairwise_similarities(model, word_list):
